gextract methylation data

gextract methylation data

gextract_meth(
  tracks,
  names = NULL,
  intervals = gintervals.all(),
  iterator = "intervs.global.seq_CG",
  d_expand = NULL,
  join_intervals = FALSE,
  extract_meth_calls = FALSE,
  avg_tracks = NULL,
  avg_tracks_names = NULL,
  annot_tracks = NULL,
  annot_tracks_names = NULL,
  min_cov = NULL,
  ...
)

Arguments

tracks

name of methylation track suffix (without ".cov" or ".meth")

names

sets the columns names in the returned value.

intervals

genomic scope for which the function is applied

iterator

track expression iterator. If "NULL" iterator is based on CpGs

d_expand

smooth the methylation signal d_expand bp from each side (optional).

join_intervals

add the intervals to the returned data frame (using left_join)

extract_meth_calls

extract also methylation calls (".meth" tracks)

avg_tracks

tracks of average methylation (for tracks that do not have ".cov" and ".meth" tracks)

avg_tracks_names

names of tracks in avg_tracks (optional)

annot_tracks

other tracks to extract

annot_tracks_names

names of annotation tracks

min_cov

minimal coverage per track. If the coverage is below this number the methylation values would be changed to NA.

...

additional arguments to gextract

Value

data frame with a column with average methylation for each track, and another column with ".cov" suffix with every track's coverage. If extract_meth_calls is TRUE - additional columns with ".meth" suffix with number of methylation calls per track.

See also